+++ ca/django_ca/management/commands/sign_cert.py
@@ -18,19 +18,16 @@
 
 import sys
 import typing
-import warnings
 from datetime import timedelta
 
 from cryptography import x509
 from cryptography.hazmat.backends import default_backend
-from cryptography.hazmat.primitives.serialization import Encoding
 
 from django.core.management.base import CommandError
 from django.core.management.base import CommandParser
 from django.utils import timezone
 
 from ... import ca_settings
-from ...deprecation import RemovedInDjangoCA120Warning
 from ...extensions import Extension
 from ...extensions import SubjectAlternativeName
 from ...management.base import BaseSignCommand
@@ -95,12 +92,6 @@ https://django-ca.readthedocs.io/en/latest/extensions.html for more information.
             metavar="FILE",
             help="The path to the certificate to sign, if ommitted, you will be be prompted.",
         )
-        self.add_format(
-            parser,
-            opts=["--csr-format"],
-            default=None,
-            help_text="Format of the CSR (DEPRECATED, format is now auto-detected).",
-        )
 
         self.add_profile(
             parser,
@@ -115,7 +106,6 @@ https://django-ca.readthedocs.io/en/latest/extensions.html for more information.
         expires: timedelta,
         watch: typing.List[str],
         password: typing.Optional[bytes],
-        encoding: typing.Optional[Encoding],
         cn_in_san: bool,
         csr_path: str,
         profile: typing.Optional[str],
@@ -126,11 +116,6 @@ https://django-ca.readthedocs.io/en/latest/extensions.html for more information.
             raise CommandError("Certificate Authority has expired.")
         if ca.revoked:
             raise CommandError("Certificate Authority is revoked.")
-        if encoding is not None:
-            warnings.warn(
-                "--csr-format option is deprecated and will be removed in django-ca 1.20.0.",
-                category=RemovedInDjangoCA120Warning,
-            )
         self.test_options(ca=ca, expires=expires, password=password, **options)
         subject = subject or Subject()